% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hospitalisation.R
\name{add_hospitalisation}
\alias{add_hospitalisation}
\title{add_hospitalisation}
\usage{
add_hospitalisation(variants_aggregated, hospitalisation_formatted)
}
\arguments{
\item{hospitalisation_formatted}{}
}
\description{
add_hospitalisation
}
