% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_variant.R
\name{enrichment_variant}
\alias{enrichment_variant}
\title{Title}
\usage{
enrichment_variant(
  data_aggregated,
  variable,
  group,
  variants,
  multiplicateur,
  time
)
}
\arguments{
\item{time}{}
}
\value{

}
\description{
Title
}
