% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_hospitalisation.R
\name{format_hospitalisation}
\alias{format_hospitalisation}
\title{disaggregate hospitalisation files}
\usage{
format_hospitalisation(hospitalisation, case_aggregated, start, end)
}
\arguments{
\item{case_aggregated}{}
}
\description{
disaggregate hospitalisation files
}
